Node registration works
Had to delay creation of i2c node. Previously it was done
in the probing of the i2c driver. During this probing, the
driver asks the device for data. Data is sent using an
smbus alert - which is handled by the smbus driver by
iterating over all children of the i2c adapter. Unfortunately,
the i2c device at time of probing has not been added as a
child to the adapter.
Thus, the creation of children to the i2c device is done
in a delayed work queue, after probing is completed.

Signed-off-by: Hans Odeberg <hans.odeberg@intel.com>